modified:   src/declarations.d.ts 	new file:   src/fetches/fetchOpportunity.tsx 	modified:   src/opportunities/components/PostsField.tsx 	modified:   src/opportunities/pages/IndividualPost.tsx
