+++ src/opportunities/pages/IndividualPost.tsx
@@ -2,50 +2,15 @@ import React, { useState } from "react";
 import JobDetails from "../components/JobDetails";
 import { useParams } from "react-router";
 import { useEffect } from "react";
-
-interface getOpportunityData {
-  "id": string;
-  "name": string;
-  "description": string;
-  "recommended_experience": string;
-  "authors": string;
-  "department": string;
-  "pay": string;
-  "credits"?: string;
-  "semester": string
-  "application_due": string;
-  "recommended_class_years": string;
-}
+import fetchOpportunity from "../../fetches/fetchOpportunity.tsx";
 
 const IndividualPost = () => {
-  const { postID } = useParams();
+  const { id } = useParams();
 
   var [details, setDetails] = useState<string | getOpportunityData>("Searching");
 
-  const fetchOpportunity = async () => {
-    // Consider moving the base URL to a configuration
-    const baseURL = `${process.env.REACT_APP_BACKEND_SERVER}`;
-    const url = `${baseURL}/getOpportunity/${postID}`;
-
-    const response = await fetch(url);
-
-    if (!response.ok) {
-      return false;
-    }
-
-    const data = await response.json();
-    console.log(data);
-    return data["data"];
-  };
-
-  async function findDetails() {
-    const data = await fetchOpportunity();
-    details = data || "Nothing found";
-    setDetails(details);
-  }
-
   useEffect(() => {
-    findDetails();
+    fetchOpportunity({setDetails, id});
   });
 
   return (
